Interop: Bring up to speed with libhb. Exposes more functions, accommodates channel layout and filter changes and allows setting CFR same as source.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4969 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrake.Interop/HandBrakeInterop/SourceData/AudioTrack.cs
@@ -47,7 +47,7 @@ namespace HandBrake.Interop.SourceData
 		/// <summary>
 		/// Gets or sets the channel layout of this Audio Track.
 		/// </summary>
-		public int ChannelLayout { get; set; }
+		public ulong ChannelLayout { get; set; }
 
 		/// <summary>
 		/// Gets or sets the frequency (in Hz) of this Audio Track
